parameters: Unify/simplify dynamic/continuous/flex parameters
Mutwo parameters that can be identified by only one value inherit from
'core_parameters.abc.SingleValueParameter'. But what if these parameters
are dynamically changing during the duration of one event (e.g.
glissando, ritardando, crescendo, ...) ? Before this patch this problem
was approached in two different ways:

    (1) The 'SingleValueParameter' Pitch has the attribute '.envelope'.
        This '.envelope' attribute can host a pitch interval trajectory
        that can be applied on the main pitch to create a glissando.
        Therefore all pitch classes have the option of an optional
        glissando. Whether this '.envelope' attribute is ignored or
        used is a decision of a given converter (or other (mutwo) code).

    (2) The 'SingleValueParameter' Tempo doesn't have the attribute
        '.envelope', but it is still possible to represent dynamic
        tempo trajectories by an additional 'TempoEnvelope'. This
        additional 'TempoEnvelope' itself isn't a 'Parameter' and
        certainly not a 'Tempo', but an 'Event' (that inherits from
        'Envelope'). Mutwo infrastructure either expects a
        'TempoEnvelope' or a 'Tempo', but never both.

This patch aims to unify the way how 'SingleValueParameter' approach
representations of continuous states. Instead of (1) or (2) another
solution is declared, that also aims to significantly simplify how
users can define dynamic and static states of
'SingleValueParameter'. The unification aims to reduce code duplications
and to make it easier to understand the 'mutwo' framework (provide only
one way that needs to be learned and apply it for everything).

Generally we could say that (1) is better than (2), because it treats
dynamic/flexible and static parameters in the same way and therefore makes
it flexible to switch between them as needed. But more precisely, the
solution (1) actually generalizes all discreet parameters to be
dynamic by giving all parameter classes an '.envelope' attribute:
therefore actual discreet values don't exist anymore. This means that
if users only want to use (the simpler) static parameters they still
carried within their code dynamic representations while they aren't
needed. Furthermore this approach has the problem that the trajectory
always needs to be relative to the identity of the parameter itself
to be generally apply-able to all parameter implementations. This mean
practically that trajectories can only be described in relative to
the central value via deltas. In case of 'Pitch' this means
we can only describe glissandi via pitch intervals that represent
the moving delta to the 'Pitch' identity, but we can't describe
glissandi by Pitch interpolations.

The unifying solution therefore is different than approach (1) or (2).
Neither do we add an '.envelope' attribute to all 'SingleValueParameter'
nor do we use external events to describe dynamic parameters. Instead,
we do add a new 'parameters.abc' class: 'FlexParameterMixin'.
In order to create a dynamic version of a parameter, the new
class can be created by inheriting from the parameter abc and the
'FlexParameterMixin'. This newly created class is both a
'Parameter' and an 'Event'. It must equally well implement the 'Parameters'
interface, so that it is usable just like any other implementation
of the given 'Parameter', but it must also behaves like an 'Envelope' so
that dynamic trajectories of the parameter can be represented and
interpreted if needed. This default version is an absolute envelope
(as this is simpler), but a relative version could also easily be
implemented if a user prefers this representation.

Details: Reasons for parameter naming as 'FlexParameter'
========================================================

I had various ideas how these parameters should be called:

1. DynamicXXX
2. ContinuousXXX
3. FlexXXX

'Dynamic': As the opposite of static (e.g. normally a parameter doesn't
change and is therefore 'static'), this seemed to be suitable. It's a
bit long and it could also mean 'having a lot of energy', which doesn't
exactly fit here. Furthermore these envelope parameters don't
necessarily need to change all the time: they could also only change for
a short time at the beginning and then stay static for the remaining
time.

'Continuous': As the opposite of discreet (e.g. normally a parameter
only has one specific discreet value). On the other hand, the envelope
parameter is also described by discreet value between which we
interpolate. Furthermore the parameter doesn't necessarily need to
change continuously: it could also keep the same for some time. It's
even longer than dynamic and with all the 'u' in the last syllable it's
easy to misspell for non-native speaker.

'Flex': As an abbreviation for 'flexible' (and therefore the opposite of
inflexible: normal parameters can only be themselves and not flexibly change),
but also as a verb to bend something: We bend the normally 'static'
parameters. This also includes the case where we only slightly change a
parameter for a short time. Furthermore it's short, easy to spell and
easy to remember. In the DAW Logic, there also exists 'Flex Time' and 'Flex
Pitch' which which pre-recorded signals can be adjusted [1]. Although
the mutwo meaning is a bit different, it also includes the same logic of
using pre-existing data ('normal' parameters) and adjusting them
(interpolating between them) to create new data.

---

[1] https://support.apple.com/guide/logicpro/flex-time-and-pitch-overview-lgcp15968647/mac