parmeters/Tempo += .seconds; converters -= TempoToBeathLengthInSeconds
This patch drops 'core_converters.TempoToBeathLengthInSeconds', but adds
the new '.seconds' property to 'core_parameters.abc.Tempo'. This new
attribute provides the same value like what the converter previously did:
it tells how long a beat is with the current tempo. Because this
calculation (from BPM to beat duration) is trivial, the encapsulation to
a specific converter isn't necessary and rather adds additional
unnecessary noise to the code. This patch therefore aims to simplify the
code and to make it more concise and readable.